fix: refactor processFields function
