+++ packages/backend/src/apps/gathersg/auth/decrypt-response.ts
@@ -5,7 +5,7 @@ import crypto from 'crypto'
 import appConfig from '@/config/app'
 import logger from '@/helpers/logger'
 
-import { HEX_ENCODED_FIELD_PREFIX } from '../common/constants'
+import { processFields } from '../common/utils'
 
 import schema from './schema'
 
@@ -31,42 +31,6 @@ function validateData(data: any, flowId: string, app: string) {
     )
   }
   return validationResult.data
-}
-
-export function processFields(fields: Record<string, any>) {
-  const processedFields: Record<string, any> = {}
-  const invalidCharRegex = /[^a-zA-Z0-9-_ ]/
-  for (const [key, value] of Object.entries(fields)) {
-    const hexKey = invalidCharRegex.test(key)
-      ? `${HEX_ENCODED_FIELD_PREFIX}${Buffer.from(key).toString('hex')}`
-      : key
-
-    // Convert primitive arrays to objects for backward compatibility
-    if (Array.isArray(value) && value.length > 0) {
-      // Check if it's an array of primitives
-      if (typeof value[0] !== 'object' || value[0] === null) {
-        // Create an object with both individual elements and the full array
-        const arrayObject: Record<string | number, any> = {}
-
-        // Add individual array elements
-        for (let i = 0; i < value.length; i++) {
-          arrayObject[i] = value[i]
-        }
-
-        // Add the full array for the new type: 'array' field
-        arrayObject._array = value
-
-        processedFields[hexKey] = arrayObject
-      } else {
-        // Array of objects - keep as is
-        processedFields[hexKey] = value
-      }
-    } else {
-      // Not an array or empty array - keep as is
-      processedFields[hexKey] = value
-    }
-  }
-  return processedFields
 }
 
 function verifySignature(signature: string, basestring: string) {
